+++ Samples/tvOS-Swift/tvOS-Swift/ContentView.swift
@@ -48,6 +48,22 @@ struct ContentView: View {
         })
     }
     
+    var oomCrashAction: () -> Void = {
+        DispatchQueue.main.async {
+            let megaByte = 1_024 * 1_024
+            let memoryPageSize = NSPageSize()
+            let memoryPages = megaByte / memoryPageSize
+            
+            while true {
+                // Allocate one MB and set one element of each memory page to something.
+                let ptr = UnsafeMutablePointer<Int8>.allocate(capacity: megaByte)
+                for i in 0..<memoryPages {
+                    ptr[i * memoryPageSize] = 40
+                }
+            }
+        }
+    }
+    
     var body: some View {
         VStack {
             Button(action: addBreadcrumbAction) {
@@ -80,8 +96,10 @@ struct ContentView: View {
                 Text("Crash")
             }
             
+            Button(action: oomCrashAction) {
+                Text("OOM Crash")
+            }
         }
-        
     }
 }